--- testdata/artifacts/package/example/0.0.2/kibana/infrastructure-ui-source/default.json
@@ -1,35 +1,0 @@
-{
-  "infrastructure-ui-source": {
-    "description": "",
-    "fields": {
-      "container": "container.id",
-      "host": "host.name",
-      "pod": "kubernetes.pod.uid",
-      "tiebreaker": "_doc",
-      "timestamp": "@timestamp"
-    },
-    "logAlias": "example-*",
-    "logColumns": [
-      {
-        "timestampColumn": {
-          "id": "5e7f964a-be8a-40d8-88d2-fbcfbdca0e2f"
-        }
-      },
-      {
-        "fieldColumn": {
-          "field": "event.dataset",
-          "id": " eb9777a8-fcd3-420e-ba7d-172fff6da7a2"
-        }
-      },
-      {
-        "messageColumn": {
-          "id": "b645d6da-824b-4723-9a2a-e8cece1645c0"
-        }
-      }
-    ],
-    "metricAlias": "example-*",
-    "name": "Default"
-  },
-  "type": "infrastructure-ui-source",
-  "updated_at": "2019-09-05T09:42:51.592Z"
-}
