--- testdata/artifacts/package/example/0.0.2/manifest.yml
@@ -1,18 +1,0 @@
-format_version: 1.0.0
-
-name: example
-title: Example
-description: This is the example integration.
-version: 0.0.2
-categories: ["logs"]
-release: beta
-
-requirement:
-  elasticsearch:
-    versions: ">=6.0.0 <=6.8.0"
-  kibana:
-    versions: ">=6.0.0"
-  metricbeat:
-    versions: ">=6.1.0"
-  filebeat:
-    versions: ">=6.2.0"
