--- testdata/artifacts/package/example/1.0.0/dataset/foo/agent/stream/stream.yml.hbs
@@ -1,1 +1,0 @@
-foo: bar
