--- testdata/artifacts/package/example/1.0.0/dataset/foo/fields/base-fields.yml
@@ -1,17 +1,0 @@
-
-- name: stream.type
-  type: constant_keyword
-  description: >
-    Stream type
-- name: stream.dataset
-  type: constant_keyword
-  description: >
-    Stream dataset.
-- name: stream.namespace
-  type: constant_keyword
-  description: >
-    Stream namespace.
-- name: "@timestamp"
-  type: date
-  description: >
-    Event timestamp.
