--- testdata/artifacts/package/example/1.0.0/dataset/foo/manifest.yml
@@ -1,17 +1,0 @@
-# This dataset has a different id then the path
-id: bar
-
-title: Foo
-
-# Needs to describe the type of this input
-type: logs
-ingest_pipeline: pipeline-entry
-
-streams:
-  - input: foo
-    vars:
-      - name: paths
-        required: true
-        description: Path to log files to be collected
-        type: text
-        multi: true
