--- testdata/artifacts/package/example/1.0.0/docs/README.md
@@ -1,3 +1,0 @@
-# Example readme
-
-This is a readme.
