--- testdata/artifacts/package/example/1.0.0/index.json
@@ -1,120 +1,0 @@
-{
-  "format_version": "1.0.0",
-  "name": "example",
-  "title": "Example Integration",
-  "version": "1.0.0",
-  "readme": "/package/example/1.0.0/docs/README.md",
-  "license": "basic",
-  "description": "This is the example integration",
-  "type": "integration",
-  "categories": [
-    "logs",
-    "metrics"
-  ],
-  "release": "ga",
-  "removable": true,
-  "requirement": {
-    "kibana": {
-      "versions": "\u003e=7.0.0"
-    },
-    "elasticsearch": {
-      "versions": "\u003e=7.0.0 \u003c=7.5.0"
-    }
-  },
-  "screenshots": [
-    {
-      "src": "/package/example/1.0.0/img/kibana-iptables.png",
-      "title": "IP Tables Overview dashboard",
-      "size": "1492x1382"
-    },
-    {
-      "src": "/package/example/1.0.0/img/kibana-iptables-ubiquity.png",
-      "title": "IP Tables Ubiquity Dashboard",
-      "size": "1492x1464",
-      "type": "image/png"
-    }
-  ],
-  "assets": [
-    "/package/example/1.0.0/manifest.yml",
-    "/package/example/1.0.0/docs/README.md",
-    "/package/example/1.0.0/img/icon.png",
-    "/package/example/1.0.0/img/kibana-envoyproxy.jpg",
-    "/package/example/1.0.0/dataset/foo/manifest.yml",
-    "/package/example/1.0.0/kibana/dashboard/0c610510-5cbd-11e9-8477-077ec9664dbd.json",
-    "/package/example/1.0.0/kibana/index-pattern/filebeat.json",
-    "/package/example/1.0.0/kibana/visualization/0a994af0-5c9d-11e9-8477-077ec9664dbd.json",
-    "/package/example/1.0.0/kibana/visualization/36f872a0-5c03-11e9-85b4-19d0072eb4f2.json",
-    "/package/example/1.0.0/kibana/visualization/38f96190-5c99-11e9-8477-077ec9664dbd.json",
-    "/package/example/1.0.0/kibana/visualization/7e4084e0-5c99-11e9-8477-077ec9664dbd.json",
-    "/package/example/1.0.0/kibana/visualization/80844540-5c97-11e9-8477-077ec9664dbd.json",
-    "/package/example/1.0.0/kibana/visualization/ab48c3f0-5ca6-11e9-8477-077ec9664dbd.json",
-    "/package/example/1.0.0/dataset/foo/fields/base-fields.yml",
-    "/package/example/1.0.0/dataset/foo/agent/stream/stream.yml.hbs",
-    "/package/example/1.0.0/dataset/foo/elasticsearch/ingest-pipeline/pipeline-entry.json",
-    "/package/example/1.0.0/dataset/foo/elasticsearch/ingest-pipeline/pipeline-http.json",
-    "/package/example/1.0.0/dataset/foo/elasticsearch/ingest-pipeline/pipeline-json.json",
-    "/package/example/1.0.0/dataset/foo/elasticsearch/ingest-pipeline/pipeline-plaintext.json",
-    "/package/example/1.0.0/dataset/foo/elasticsearch/ingest-pipeline/pipeline-tcp.json"
-  ],
-  "datasets": [
-    {
-      "id": "bar",
-      "title": "Foo",
-      "release": "experimental",
-      "type": "logs",
-      "ingest_pipeline": "pipeline-entry",
-      "streams": [
-        {
-          "input": "foo",
-          "vars": [
-            {
-              "name": "paths",
-              "type": "text",
-              "description": "Path to log files to be collected",
-              "multi": true,
-              "required": true,
-              "show_user": false
-            }
-          ],
-          "enabled": true
-        }
-      ],
-      "package": "example",
-      "path": "foo"
-    }
-  ],
-  "datasources": [
-    {
-      "name": "logs",
-      "title": "Logs datasource",
-      "description": "Datasource for your log files.",
-      "inputs": [
-        {
-          "type": "foo",
-          "streams": [
-            {
-              "input": "foo",
-              "vars": [
-                {
-                  "name": "paths",
-                  "type": "text",
-                  "description": "Path to log files to be collected",
-                  "multi": true,
-                  "required": true,
-                  "show_user": false
-                }
-              ],
-              "dataset": "bar",
-              "template_path": "stream.yml.hbs",
-              "template": "foo: bar\n",
-              "enabled": true
-            }
-          ]
-        }
-      ],
-      "multiple": true
-    }
-  ],
-  "download": "/epr/example/example-1.0.0.tar.gz",
-  "path": "/package/example/1.0.0"
-}
