--- testdata/artifacts/package/example/1.0.0/manifest.yml
@@ -1,35 +1,0 @@
-format_version: 1.0.0
-
-name: example
-description: This is the example integration
-version: 1.0.0
-title: Example Integration
-categories: ["logs", "metrics"]
-type: integration
-release: ga
-
-requirement:
-  elasticsearch:
-    versions: ">=7.0.0 <=7.5.0"
-  kibana:
-    versions: ">=7.0.0"
-  metricbeat:
-    versions: ">=7.1.0"
-  filebeat:
-    versions: ">=7.2.0"
-
-screenshots:
-  - src: /img/kibana-iptables.png
-    title: IP Tables Overview dashboard
-    size: 1492x1382
-  - src: /img/kibana-iptables-ubiquity.png
-    title: IP Tables Ubiquity Dashboard
-    size: 1492x1464
-    type: image/png
-
-datasources:
-  - name: logs
-    title: Logs datasource
-    description: Datasource for your log files.
-    inputs:
-      - type: foo
